Don't remove the dist-packages directory when creating the dist tarball
In the upstream version of setup.py, this directory gets checked for
build commands such as 'sdist' right before creating the distribution
even in cases where --static is not passed as an argument, and one out
of two possible actions are taken in that directory is NOT empty:

 1. If running from sources or a build command (e.g. sdist), the
    directory gets simply removed and its bits discarded, as it's
    assumed that we're building the "dynamic" version of KA Lite,
    which uses system's libraries.

 2. Otherwise, DIST_NAME changes to 'ka-lite-static' as if we were
    building with --static, even though this code gets only executed
    when 'not STATIC_BUILD', which seems a bit contradictory.

This behaviour is giving us trouble when creating the debian sources
in Jenkins from the git repository, since we are not building with
--static (as we don't want to use pip to automatically download the
libraries to be bundled, but the ones already in the repo).

However, at the same time we don't really want to delete the contents
of the dist-packages repository, since that would mean that the result
debian sources would miss the bundled python modules we have in the
git repository (which was based on the upstream debian package).

So, the best solution seems to be to simply remove that strange code
path that would remove the dist-packages directory, if it exists,
before running sdist, so that both KA Lite and its required modules
get bundled, the package properly built and the flatpak running.

https://phabricator.endlessm.com/T4489